pyproject.toml
[tool.poetry]
name = "boaserver"
version = "0.1.0"
description = "AI-Powered Oncall Support Assistant"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "boaserver"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.0"
uvicorn = "^0.22.0"
pydantic = "^1.10.7"
aiohttp = "^3.8.4"
python-dotenv = "^1.0.0"
google-generativeai = "^0.2.0"
httpx = "^0.24.0"
python-multipart = "^0.0.6"
tenacity = "^8.2.2"
websockets = "^11.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
ruff = "^0.0.267"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "B", "I"]
